pyproject.toml
[project]
name = "chatbot-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MianAnwar", email = "miananwar.developer@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.99.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
chatbot-api = "chatbot_api:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
